requirements.txt
# requirements.txt

allure-pytest==2.13.5
loguru==0.7.2
cryptography==42.0.8
DingtalkChatbot==1.5.7
Faker==25.8.0
Jinja2==3.1.4
mitmproxy==10.3.1
openpyxl==3.1.4
pandas==2.2.2
pytest==8.2.2
pytest-xdist==3.6.1
PyMySQL==1.1.1
pyparsing==3.1.2
pyperclip==1.8.2
pydantic==2.7.4
PyYAML==6.0.1
redis==5.0.6
requests==2.32.3
requests-toolbelt==1.0.0
ruamel.yaml==0.18.6
urllib3==2.2.2
jsonpath==0.82.2
xlrd==2.0.1
xlwt==1.3.0
